[core][Android] Rename `AsyncFunction` -> `AsyncFunctionComponent`
